+++ src/main/java/com/landawn/streamex/Triple.java
@@ -1,18 +1,19 @@
 /*
- * Copyright 2015, 2016 Tagir Valeev
+ * Copyright (c) 2015, Haiyang Li.
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
- *     http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package com.landawn.streamex;
 
 import java.util.Arrays;
@@ -20,22 +21,31 @@ import java.util.Collection;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.function.BiPredicate;
 import java.util.function.Consumer;
-import java.util.function.DoubleConsumer;
-import java.util.function.IntConsumer;
-import java.util.function.LongConsumer;
+import java.util.function.Function;
+import java.util.function.Predicate;
 
 import com.landawn.streamex.function.QuadPredicate;
+import com.landawn.streamex.function.TriConsumer;
+import com.landawn.streamex.function.TriFunction;
+import com.landawn.streamex.function.TriPredicate;
 
 /**
- * @author haiyangl
+ * 
+ * @since 0.8
+ * 
+ * @author Haiyang Li
  *
+ * @param <L>
+ * @param <M>
+ * @param <R>
  */
 public final class Triple<L, M, R> {
-    private volatile L left;
-    private volatile M middle;
-    private volatile R right;
+    public volatile L left;
+    public volatile M middle;
+    public volatile R right;
 
     public Triple() {
     }
@@ -62,13 +72,13 @@ public final class Triple<L, M, R> {
         }
     }
 
+    @SuppressWarnings("unchecked")
     public static <T> Triple<T, T, T> from(Collection<? extends T> c) {
         if (c == null || c.size() == 0) {
             return new Triple<>(null, null, null);
         }
 
-        @SuppressWarnings("unchecked")
-        List<T> list = c instanceof List ? (List<T>) c : null;
+        final List<T> list = c instanceof List ? (List<T>) c : null;
 
         if (c.size() == 1) {
             if (list != null) {
@@ -307,6 +317,30 @@ public final class Triple<L, M, R> {
         objComsumer.accept(right);
     }
 
+    public void accept(final TriConsumer<? super L, ? super M, ? super R> action) {
+        action.accept(left, middle, right);
+    }
+
+    public void accept(final Consumer<Triple<L, M, R>> action) {
+        action.accept(this);
+    }
+
+    public <U> U map(final TriFunction<? super L, ? super M, ? super R, U> mapper) {
+        return mapper.apply(left, middle, right);
+    }
+
+    public <U> U map(final Function<Triple<L, M, R>, U> mapper) {
+        return mapper.apply(this);
+    }
+
+    public Optional<Triple<L, M, R>> filter(final TriPredicate<? super L, ? super M, ? super R> predicate) {
+        return predicate.test(left, middle, right) ? Optional.of(this) : Optional.<Triple<L, M, R>> empty();
+    }
+
+    public Optional<Triple<L, M, R>> filter(final Predicate<Triple<L, M, R>> predicate) {
+        return predicate.test(this) ? Optional.of(this) : Optional.<Triple<L, M, R>> empty();
+    }
+
     public StreamEx<Triple<L, M, R>> stream() {
         return StreamEx.of(this);
     }
@@ -341,221 +375,5 @@ public final class Triple<L, M, R> {
     @Override
     public String toString() {
         return "[" + Objects.toString(left) + ", " + Objects.toString(middle) + ", " + Objects.toString(right) + "]";
-    }
-
-    public static final class IntTriple {
-        public final int _1;
-        public final int _2;
-        public final int _3;
-
-        private IntTriple(int _1, int _2, int _3) {
-            this._1 = _1;
-            this._2 = _2;
-            this._3 = _3;
-        }
-
-        public static IntTriple of(int _1, int _2, int _3) {
-            return new IntTriple(_1, _2, _3);
-        }
-
-        public int min() {
-            return Math.min(_1 <= _2 ? _1 : _2, _3);
-        }
-
-        public int max() {
-            return Math.max(_1 >= _2 ? _1 : _2, _3);
-        }
-
-        public int sum() {
-            return _1 + _2 + _3;
-        }
-
-        public double average() {
-            return sum() / 3;
-        }
-
-        public IntTriple reversed() {
-            return new IntTriple(_3, _2, _1);
-        }
-
-        public int[] toArray() {
-            return new int[] { _1, _2, _3 };
-        }
-
-        public void forEach(IntConsumer comsumer) {
-            comsumer.accept(this._1);
-            comsumer.accept(this._2);
-            comsumer.accept(this._3);
-        }
-
-        public StreamEx<IntTriple> stream() {
-            return StreamEx.of(this);
-        }
-
-        @Override
-        public int hashCode() {
-            return (31 * (31 * _1 + this._2)) + _3;
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (this == obj) {
-                return true;
-            } else if (!(obj instanceof IntTriple)) {
-                return false;
-            } else {
-                IntTriple other = (IntTriple) obj;
-                return this._1 == other._1 && this._2 == other._2 && this._3 == other._3;
-            }
-        }
-
-        @Override
-        public String toString() {
-            return "[" + this._1 + ", " + this._2 + ", " + this._3 + "]";
-        }
-    }
-
-    public static final class LongTriple {
-        public final long _1;
-        public final long _2;
-        public final long _3;
-
-        private LongTriple(long _1, long _2, long _3) {
-            this._1 = _1;
-            this._2 = _2;
-            this._3 = _3;
-        }
-
-        public static LongTriple of(long _1, long _2, long _3) {
-            return new LongTriple(_1, _2, _3);
-        }
-
-        public long min() {
-            return Math.min(_1 <= _2 ? _1 : _2, _3);
-        }
-
-        public long max() {
-            return Math.max(_1 >= _2 ? _1 : _2, _3);
-        }
-
-        public long sum() {
-            return _1 + _2 + _3;
-        }
-
-        public double average() {
-            return sum() / 3;
-        }
-
-        public LongTriple reversed() {
-            return new LongTriple(_3, _2, _1);
-        }
-
-        public long[] toArray() {
-            return new long[] { _1, _2, _3 };
-        }
-
-        public void forEach(LongConsumer comsumer) {
-            comsumer.accept(this._1);
-            comsumer.accept(this._2);
-            comsumer.accept(this._3);
-        }
-
-        public StreamEx<LongTriple> stream() {
-            return StreamEx.of(this);
-        }
-
-        @Override
-        public int hashCode() {
-            return (int) ((31 * (31 * _1 + this._2)) + _3);
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (this == obj) {
-                return true;
-            } else if (!(obj instanceof LongTriple)) {
-                return false;
-            } else {
-                LongTriple other = (LongTriple) obj;
-                return this._1 == other._1 && this._2 == other._2 && this._3 == other._3;
-            }
-        }
-
-        @Override
-        public String toString() {
-            return "[" + this._1 + ", " + this._2 + ", " + this._3 + "]";
-        }
-    }
-
-    public static final class DoubleTriple {
-        public final double _1;
-        public final double _2;
-        public final double _3;
-
-        private DoubleTriple(double _1, double _2, double _3) {
-            this._1 = _1;
-            this._2 = _2;
-            this._3 = _3;
-        }
-
-        public static DoubleTriple of(double _1, double _2, double _3) {
-            return new DoubleTriple(_1, _2, _3);
-        }
-
-        public double min() {
-            return Math.min(_1 <= _2 ? _1 : _2, _3);
-        }
-
-        public double max() {
-            return Math.max(_1 >= _2 ? _1 : _2, _3);
-        }
-
-        public double sum() {
-            return _1 + _2 + _3;
-        }
-
-        public double average() {
-            return sum() / 3;
-        }
-
-        public DoubleTriple reversed() {
-            return new DoubleTriple(_3, _2, _1);
-        }
-
-        public double[] toArray() {
-            return new double[] { _1, _2, _3 };
-        }
-
-        public void forEach(DoubleConsumer comsumer) {
-            comsumer.accept(this._1);
-            comsumer.accept(this._2);
-            comsumer.accept(this._3);
-        }
-
-        public StreamEx<DoubleTriple> stream() {
-            return StreamEx.of(this);
-        }
-
-        @Override
-        public int hashCode() {
-            return (int) ((31 * (31 * _1 + this._2)) + _3);
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (this == obj) {
-                return true;
-            } else if (!(obj instanceof DoubleTriple)) {
-                return false;
-            } else {
-                DoubleTriple other = (DoubleTriple) obj;
-                return this._1 == other._1 && this._2 == other._2 && this._3 == other._3;
-            }
-        }
-
-        @Override
-        public String toString() {
-            return "[" + this._1 + ", " + this._2 + ", " + this._3 + "]";
-        }
     }
 }